Version 1.2.0: Backface culling!
It now calculates the normal of every face and it decides whether it should render that face or not by comparing the distance from the normal of the face to the camera with the distance from the center of the face to the camera. Some faces have inverted normals so I hardcoded them. (don't do that yourself)